adkpy/agents/requirements.txt
# ArangoDB Client Integration Requirements
python-arango>=7.7.0
asyncio-pool>=0.6.0
tenacity>=8.2.0

# Existing ADK requirements
google-genai>=0.7.0
google-adk>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Additional utilities
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
prometheus-client>=0.19.0  # For metrics collection